+++ lib/reform/form/active_model.rb
@@ -17,7 +17,6 @@ module Reform::Form::ActiveModel
     end
 
     def property(name, options={})
-puts "delegate: #{options[:on]}"
       delegate options[:on] => :model
       super
     end