+++ CHANGES.md
@@ -1,3 +1,8 @@
+h3. 0.2.0
+
+* `Form.new` now accepts one argument, only: the model/composition. If you want to create your own representer, inject it by overriding `Form#mapper`. Note that this won't create property accessors for you.
+* You can now nest forms. Note that this currently works only 1-level deep.
+
 h3. 0.1.3
 
 * Altered `reform/rails` to conditionally load `ActiveRecord` code and created `reform/active_record`.